requirements.txt
graphviz
networkx
openpyxl
pandas
plotly
psycopg2-binary
pygments
requests
sqlparse
streamlit